requirements.txt
Flask==3.0.0
openai==1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
hypothesis==6.88.1